requirements.txt
langchain
langchain_community
ipykernel
python-dotenv
langchain-openai
beautifulsoup4
faiss-cpu
streamlit
langchain-ollama